Type ordering tables and generalize battle GPU packets

+++ src/wldcore/wldcore_gfx_draw_world_frame.c
@@ -17,8 +17,8 @@ void wldcore_gfx_draw_world_frame(void) {
     wldcore_switch_to_stack((void*)0x1F8003FC);
     wldcore_gfx_draw_context_value_display(&g_wldcore_gfx_ordering_tables[g_active_graphics_buffer_index]);
     wldcore_gfx_draw_calendar_date(&g_wldcore_gfx_ordering_tables[g_active_graphics_buffer_index]);
-    wldcore_proposition_submit_marker_request((s32)&g_wldcore_gfx_ordering_tables[g_active_graphics_buffer_index]);
-    wldcore_dispatch_display_object_list((s32)&g_wldcore_gfx_ordering_tables[g_active_graphics_buffer_index],
+    wldcore_proposition_submit_marker_request(&g_wldcore_gfx_ordering_tables[g_active_graphics_buffer_index]);
+    wldcore_dispatch_display_object_list(&g_wldcore_gfx_ordering_tables[g_active_graphics_buffer_index],
         (wldcore_display_object_t**)g_wldcore_window_render_object_queue, g_wldcore_window_render_object_count);
     wldcore_fade_draw_screen_overlays();
     wldcore_restore_previous_stack();
